Fix attribute regex matching
It used to work well but since using Swift 3, the attributeRegex only matches css attributes without values. For example, "a[rel]" is correctly interpreted as ".//a[@rel]" but "a[rel='next']" isn't matched so it returns only ".//a", therefore catching all anchors in the HTML document. Because you can put quite anything as a value of an attribute, matching any character inside the brackets seems to be the right solution.